# api_server/routes/swagger-ui-dist/swagger.json
{
  "swagger": "2.0",
...
  "host": "localhost:8080",
  "basePath": "/API",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/Sources": {
      "get": {
        "tags": [
          "Sources"
        ],
        "summary": "Get all known fake news sources from the database",
        "description": "Get request to retrieve all known fake news sources from the database",
        "operationId": "",
        "produces": "application/json",
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/GET Source"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sources"
        ],
        "summary": "Adds a new fake news source to the database",
        "description": "Post request to add a new fake news source to the database",
        "operationId": "",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": "application/json",
        "parameters": [
          {
            "name": "statement",
            "in": "formData",
            "description": "Fact that is to be added",
            "required": true,
            "type": "string"
          },
          {
            "name": "popularity",
            "in": "formData",
            "description": "Popularity rating",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/POST Source"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/Sources/{SourceId}": {
      "get": {
        "tags": [
          "Sources"
        ],
        "summary": "Get a specific known fake news source from the database",
        "description": "Get request to retrieve a specific known fake news sources from the database",
        "operationId": "",
        "produces": "application/json",
        "parameters": [
          {
            "name": "SourceId",
            "in": "path",
            "description": "ID of source to be fetched",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/GET Specific Source"
            }
          },
          "404": {
            "schema": {
              "$ref": "#/definitions/GET Specific Source 404"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sources"
        ],
        "summary": "Removes a specific known fake news source, based off of it's ID, from the database",
        "description": "Delete request to remove a specific known fake news source from the database",
        "operationId": "",
        "produces": "application/json",
        "parameters": [
          {
            "name": "SourceId",
            "in": "path",
            "description": "ID of known fake news source to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/DELETE Specific Source"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Sources"
        ],
        "summary": "Update a specific known fake news source",
        "description": "Put request to update a specific known fake news source",
        "operationId": "",
        "produces": "application/json",
        "parameters": [
          {
            "name": "SourceId",
            "in": "path",
            "description": "ID of source to be updated",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "New name for the known fake news source",
            "required": false,
            "type": "string"
          },
          {
            "name": "tld",
            "in": "formData",
            "description": "New website address for the known fake news source",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/PUT Specific Source"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "Error": {
      "type": "object",
      "properties": {
        "response": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Server Error"
            },
            "success": {
              "type": "boolean",
              "example": false
            }
          }
        }
      }
    },
...
  }
}